ADM-733[backend]feat: add test to catch exception

+++ backend/src/main/java/heartbeat/handler/base/AsyncDataBaseHandler.java
@@ -2,7 +2,6 @@ package heartbeat.handler.base;
 
 import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
-import heartbeat.exception.FileIOException;
 import heartbeat.exception.GenerateReportException;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.util.ObjectUtils;
@@ -14,6 +13,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.util.Objects;
 
 import static heartbeat.handler.base.FIleType.METRICS_DATA_COMPLETED;
 import static heartbeat.service.report.scheduler.DeleteExpireCSVScheduler.EXPORT_CSV_VALIDITY_TIME;
@@ -25,15 +25,16 @@ public class AsyncDataBaseHandler {
 
 	public static final String SUFFIX_TMP = ".tmp";
 
-	protected void createDirToConvertData(FIleType fIleType) {
+	private void createDirToConvertData(FIleType fIleType) {
 		File directory = new File(OUTPUT_FILE_PATH + fIleType.getType());
 		boolean isCreateSucceed = directory.mkdirs();
 		String message = isCreateSucceed ? String.format("Successfully create %s directory", fIleType.getType())
-				: String.format("Failed to create %s directory because it already exist", fIleType.getType());
+				: String.format("Failed create %s directory because it already exist", fIleType.getType());
 		log.info(message);
 	}
 
 	protected void creatFileByType(FIleType fIleType, String fileId, String json) {
+		createDirToConvertData(fIleType);
 		String fileName = OUTPUT_FILE_PATH + fIleType.getPath() + fileId;
 		String tmpFileName = OUTPUT_FILE_PATH + fIleType.getPath() + fileId + SUFFIX_TMP;
 		log.info("Start to write file type: {}, file name: {}", fIleType.getType(), fileName);
@@ -41,11 +42,11 @@ public class AsyncDataBaseHandler {
 			writer.write(json);
 			Files.move(Path.of(tmpFileName), Path.of(fileName), StandardCopyOption.ATOMIC_MOVE,
 					StandardCopyOption.REPLACE_EXISTING);
-			log.info("Successfully write file type: {}, file name: {}", fIleType.getType(), fileName);
+			log.info("Successfully write file type: {}, file name: {}", fIleType.getType(), fileId);
 		}
-		catch (IOException e) {
-			log.error("Failed to write file type: {}, file name: {}, reason: {}", fIleType.getType(), fileName, e);
-			throw new FileIOException(e);
+		catch (IOException | RuntimeException e) {
+			log.error("Failed write file type: {}, file name: {}, reason: {}", fIleType.getType(), fileId, e);
+			throw new GenerateReportException("Failed write " + fIleType.getType() + " file " + fileId);
 		}
 	}
 
@@ -55,23 +56,29 @@ public class AsyncDataBaseHandler {
 			try (JsonReader reader = new JsonReader(new FileReader(fileName))) {
 				return new Gson().fromJson(reader, classType);
 			}
-			catch (IOException e) {
-				throw new GenerateReportException("Failed to convert to report response");
+			catch (IOException | RuntimeException e) {
+				log.error("Failed read file type: {}, file name: {}, reason: {}", fIleType.getType(), fileId, e);
+				throw new GenerateReportException("Failed read " + fIleType.getType() + " file " + fileId);
 			}
 		}
 		return null;
 	}
 
 	protected <T> T readAndRemoveFileByType(FIleType fIleType, String fileId, Class<T> classType) {
-		T t = readFileByType(fIleType, fileId, classType);
 		String fileName = OUTPUT_FILE_PATH + fIleType.getPath() + fileId;
+		log.info("Start to remove file type: {}, file name: {}", fIleType.getType(), fileId);
 		try {
-			Files.delete(Path.of(fileName));
+			T t = readFileByType(fIleType, fileId, classType);
+			if (Objects.nonNull(t)) {
+				Files.delete(Path.of(fileName));
+			}
+			log.info("Successfully remove file type: {}, file name: {}", fIleType.getType(), fileId);
+			return t;
 		}
-		catch (IOException e) {
-			throw new RuntimeException(e);
+		catch (IOException | RuntimeException e) {
+			log.info("Failed remove file type: {}, file name: {}", fIleType.getType(), fileId);
+			throw new GenerateReportException("Failed remove " + fIleType.getType() + " file " + fileId);
 		}
-		return t;
 	}
 
 	protected void deleteExpireFileByType(FIleType fIleType, long currentTimeStamp) {
